CI 수정 및 test 분리
