+++ .github/workflows/CI.yml
@@ -25,5 +25,6 @@ jobs:
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test
         run: |
+          python -m tests.utils_test
           python -m tests.__init__
           python -m server_test
